feat: asset bundling
