+++ installer/.env.development
@@ -1,0 +1,3 @@
+PORT=3002
+BROWSER=none
+REACT_APP_BEE_DESKTOP_URL=http://localhost:3000
