--- installer/src/Box.jsx
@@ -1,6 +1,0 @@
-export function Box({ children, mb }) {
-    const style = {
-        marginBottom: mb ? mb * 8 : undefined
-    }
-    return <div style={style}>{children}</div>
-}
