--- installer/src/Center.jsx
@@ -1,3 +1,0 @@
-export function Center({ children }) {
-    return <div className="center">{children}</div>
-}
