+++ installer/src/Center.tsx
@@ -1,0 +1,5 @@
+import { ReactElement } from 'react'
+
+export function Center({ children }: { children: ReactElement[] }) {
+  return <div className="center">{children}</div>
+}
