+++ installer/src/index.tsx
