+++ package.json
@@ -9,11 +9,15 @@
   "description": "Desktop client for running and managing bee",
   "main": "dist/desktop/src/index.js",
   "scripts": {
-    "start": "cross-env-shell NODE_ENV=development \"npm run clean && tsc && electron-forge start\"",
-    "package": "npm run clean && tsc && electron-forge package",
-    "make": "npm run clean && tsc && electron-forge make",
-    "publish": "npm run clean && tsc && electron-forge publish",
-    "prepareDist": "npm run clean && tsc",
+    "start": "cross-env-shell NODE_ENV=development \"npm run clean && npm run build && electron-forge start\"",
+    "package": "npm run clean && npm run build && electron-forge package",
+    "make": "npm run clean && npm run build && electron-forge make",
+    "publish": "npm run clean && npm run build && electron-forge publish",
+    "build:installer": "cd installer && npm run build && cd ..",
+    "build:desktop": "tsc",
+    "build": "npm run build:desktop && npm run build:installer && npm run copy:installer",
+    "copy:installer": "cpy . '../../dist/installer/' --cwd=installer/build",
+    "open:installer": "node ./installer.mjs",
     "clean": "rimraf dist out",
     "depcheck": "depcheck .",
     "lint": "eslint --fix \"src/**/*.ts\" && prettier --write \"src/**/*.ts\"",
@@ -100,6 +104,7 @@
     "@types/uuid": "^8.3.4",
     "@typescript-eslint/eslint-plugin": "^5.17.0",
     "@typescript-eslint/parser": "^5.17.0",
+    "cpy-cli": "^4.1.0",
     "cross-env": "^7.0.3",
     "depcheck": "^1.4.3",
     "electron": "18.0.1",
@@ -116,6 +121,7 @@
     "typescript": "~4.5.4"
   },
   "dependencies": {
+    "@ethersphere/bee-dashboard": "^0.17.0",
     "@koa/router": "^10.1.1",
     "@sentry/electron": "^3.0.7",
     "cross-zip": "^4.0.0",
